# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Documentación de la API de CentralizadorBackEnd",
    "version": "1.0.0",
    "contact": {
      "name": "Equipo de Desarrollo",
      "email": "devteam@example.com"
    },
    "description": "API REST para la gestión de datos de la plataforma CogniCare, incluyendo estudiantes, entrenadores y entrenamientos cognitivos.\n\nAutenticación\n\n **IMPORTANTE:** La mayoría de los endpoints requieren un token de autenticación (JWT Bearer).\n\nPara acceder a los endpoints protegidos, sigue estos pasos:\n1. **Obtén tu token:** Realiza una petición `POST` a `/api/login/admin` o `/api/login/entrenador` con tus credenciales.\n2. **Autoriza tus peticiones:** Haz clic en el botón **\"Authorize\"** (arriba a la derecha) y pega tu token en el campo `bearerAuth` con el formato `Bearer <TU_TOKEN_JWT>`.\n\nRoles y Permisos\n\nLa API cuenta con dos roles principales con diferentes niveles de acceso:\n\n### **Admin**\n- Gestión completa de entrenadores (CRUD).\n- Visualización de todos los informes de entrenamiento.\n- Acceso total a las utilidades del sistema.\n\n### **Entrenador**\n- Gestión de estudiantes y sus entrenamientos cognitivos.\n- Creación y seguimiento de sesiones de entrenamiento.\n- Visualización de informes de los estudiantes asignados."
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Servidor de Desarrollo Local"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token JWT obtenido al hacer login. Formato: `Bearer <tu_token>`"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Mensaje de error."
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "msg": {
                  "type": "string",
                  "description": "Mensaje de validación del error."
                },
                "param": {
                  "type": "string",
                  "description": "Parámetro asociado al error."
                },
                "location": {
                  "type": "string",
                  "description": "Ubicación del error (e.g., \"body\")."
                }
              }
            }
          }
        }
      },
...
    }
  },
  "paths": {
    "/api/admin/entrenadores": {
      "post": {
        "summary": "Registra un nuevo entrenador en el sistema.",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "siglaTipoDocumento",
                  "numeroDocumento",
                  "nombres",
                  "apellidos",
                  "correo",
                  "nombresFacultades"
                ],
                "properties": {
                  "siglaTipoDocumento": {
                    "type": "string",
                    "description": "Sigla del tipo de documento del entrenador (ej. \"CC\", \"TI\").",
                    "example": "CC"
                  },
                  "numeroDocumento": {
                    "type": "string",
                    "description": "Número de documento único del entrenador.",
                    "example": "1012345678"
                  },
                  "nombres": {
                    "type": "string",
                    "description": "Nombres del entrenador.",
                    "example": "Carlos Andres"
                  },
                  "apellidos": {
                    "type": "string",
                    "description": "Apellidos del entrenador.",
                    "example": "Ramirez Gomez"
                  },
                  "correo": {
                    "type": "string",
                    "format": "email",
                    "description": "Correo electrónico único del entrenador.",
                    "example": "carlos.ramirez@example.com"
                  },
                  "nombresFacultades": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "description": "Lista con los nombres de las facultades a las que se asignará. Debe contener al menos una.",
                    "example": [
                      "Ingenierías",
                      "Ciencias Sociales"
                    ]
                  },
                  "fechaFin": {
                    "type": "string",
                    "format": "date",
                    "description": "Opcional. Fecha de finalización del contrato. Por defecto, 6 meses desde hoy.",
                    "example": "2026-01-31"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Entrenador registrado exitosamente."
          },
          "400": {
            "$ref": "#/components/schemas/Error"
          },
          "401": {
            "$ref": "#/components/schemas/Error"
          },
          "403": {
            "$ref": "#/components/schemas/Error"
          },
          "409": {
            "description": "Conflicto, el entrenador ya existe (documento o correo duplicado).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/schemas/Error"
          }
        }
      },
      "get": {
        "summary": "Obtiene la lista de todos los entrenadores.",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de entrenadores obtenida exitosamente.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Entrenadores obtenidos exitosamente."
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "idEntrenador": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "tipoDocumentoEntrenador": {
                            "type": "string"
                          },
                          "numeroDocumentoEntrenador": {
                            "type": "string"
                          },
                          "nombresEntrenador": {
                            "type": "string"
                          },
                          "apellidosEntrenador": {
                            "type": "string"
                          },
                          "correoEntrenador": {
                            "type": "string",
                            "format": "email"
                          },
                          "fechaInicioContrato": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "fechaFinContrato": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "estadoContrato": {
                            "type": "string",
                            "example": "Activo"
                          },
                          "facultadesAsignadas": {
                            "type": "array",
                            "items": {
                              "type": "string"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/schemas/Error"
          },
          "403": {
            "$ref": "#/components/schemas/Error"
          },
          "500": {
            "$ref": "#/components/schemas/Error"
          }
        }
      }
    },
    "/api/admin/entrenadores/{entrenadorID}": {
      "put": {
        "summary": "Actualiza la información de un entrenador existente.",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "entrenadorID",
            "required": true,
            "description": "ID del entrenador (UUID) a modificar.",
            "schema": {
              "type": "string",
              "format": "uuid",
              "example": "b5f8db27-c5e0-4764-a7c8-9f7eb42302a1"
            }
          }
        ],
        "requestBody": {
          "description": "Objeto con los campos a actualizar. Solo es necesario incluir los campos que se desean cambiar.",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "pNuevosNombres": {
                    "type": "string",
                    "description": "Los nuevos nombres del entrenador.",
                    "example": "Carlos Alberto"
                  },
                  "pNuevosApellidos": {
                    "type": "string",
                    "description": "Los nuevos apellidos del entrenador.",
                    "example": "Gomez Ramirez"
                  },
                  "pNuevoCorreo": {
                    "type": "string",
                    "format": "email",
                    "description": "El nuevo correo electrónico del entrenador.",
                    "example": "c.ramirez@example.com"
                  },
                  "pNuevaFechaFin": {
                    "type": "string",
                    "format": "date",
                    "description": "La nueva fecha de finalización del contrato.",
                    "example": "2027-12-31"
                  },
                  "pNuevosNombresFacultades": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "description": "Reemplazo completo de la lista de facultades del entrenador.",
                    "example": [
                      "Ingenierías"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Entrenador actualizado exitosamente."
          },
          "400": {
            "$ref": "#/components/schemas/Error"
          },
          "401": {
            "$ref": "#/components/schemas/Error"
          },
          "403": {
            "$ref": "#/components/schemas/Error"
          },
          "404": {
            "description": "Entrenador no encontrado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/schemas/Error"
          }
        }
      }
    },
    "/api/admin/entrenadores/{entrenadorID}/desactivar": {
      "put": {
        "summary": "Desactiva un entrenador existente.",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "entrenadorID",
            "required": true,
            "description": "ID del entrenador (UUID) a desactivar.",
            "schema": {
              "type": "string",
              "format": "uuid",
              "example": "a1b2c3d4-e5f6-7890-1234-567890abcdef"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Entrenador desactivado exitosamente.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Entrenador desactivado exitosamente."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/schemas/Error"
          },
          "401": {
            "$ref": "#/components/schemas/Error"
          },
          "403": {
            "$ref": "#/components/schemas/Error"
          },
          "404": {
            "description": "Entrenador no encontrado.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "El entrenador con ID <UUID> no fue encontrado."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/schemas/Error"
          }
        }
      }
    },
    "/api/admin/informes/entrenamientos": {
      "get": {
        "summary": "Obtiene la lista de todos los entrenamientos para la vista de informes del administrador.",
        "tags": [
          "InformesAdmin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de entrenamientos obtenida exitosamente.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Lista de entrenamientos obtenida exitosamente."
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "entrenamiento_id": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "nombre_estudiante": {
                            "type": "string"
                          },
                          "apellido_estudiante": {
                            "type": "string"
                          },
                          "numero_documento_estudiante": {
                            "type": "string"
                          },
                          "tipo_documento_estudiante_sigla": {
                            "type": "string"
                          },
                          "fecha_inicio": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "fecha_fin": {
                            "type": "string",
                            "format": "date-time",
                            "nullable": true
                          },
                          "estado": {
                            "type": "string"
                          },
                          "nombre_entrenador": {
                            "type": "string"
                          },
                          "apellido_entrenador": {
                            "type": "string"
                          },
                          "facultad_nombre": {
                            "type": "string"
                          },
                          "programa_nombre": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/schemas/Error"
          },
          "403": {
            "$ref": "#/components/schemas/Error"
          },
          "500": {
            "$ref": "#/components/schemas/Error"
          }
        }
      }
    },
...
  },
...
}